Move some utility functions to modality-specific modules
