+++ tests/models/decoder_only/vision_language/test_awq.py
@@ -3,7 +3,7 @@ from typing import List, Optional, Type
 import pytest
 import torch
 
-from vllm.multimodal.utils import rescale_image_size
+from vllm.multimodal.image import rescale_image_size
 
 from ....conftest import IMAGE_ASSETS, VllmRunner, _ImageAssets
 from ...utils import check_logprobs_close
